Refactored story components and related routing
